Chore: boxed primitives, cleanup, mailing and errors
* fix: boxed primitives in services and DAOs

* fix: boxed primitives in controllers

* fix: boxed primitives Querys/Forms

* chore: cleanup in controllers

* chore: cleanup and boxed primitives in models

* chore: cleanup Collections.emptyList()

* Merge branch 'chore/rest' into chore/boxed-primitives

# Conflicts:
#	webapp/src/main/java/ar/edu/itba/paw/webapp/controller/UserController.java

* fix: merging and invalid avatar

* fix: games and mailing

* fix: errors frontend


Approved-by: Federico Shih

+++ persistence/src/main/java/ar/edu/itba/paw/persistence/ReportHibernateDao.java
@@ -95,7 +95,7 @@ public class ReportHibernateDao implements ReportDao, PaginationDao<ReportFilter
         QueryBuilder queryBuilder = getQueryBuilderFromFilter(filter);
         Query nativeQuery = em.createNativeQuery("SELECT id from reports " + queryBuilder.toQuery());
         prepareParametersForNativeQuery(queryBuilder, nativeQuery);
-        nativeQuery.setFirstResult(page.getOffset().intValue());
+        nativeQuery.setFirstResult(page.getOffset());
         nativeQuery.setMaxResults(page.getPageSize());
 
         @SuppressWarnings("unchecked")
